\docType{data}
\name{combined_areas}
\alias{combined_areas}
\title{Combined Statistical Areas (CSAs)}
\format{A data frame with 166 rows and 2 variables.}
\usage{
data(combined_areas)
}
\description{
The U.S. Census Bureau groups counties into CSAs primarily
based on county population. NOTE: Not all counties are
members of a CSA. For a detailed description, Wikipedia has
an excellent discussion:
\url{http://en.wikipedia.org/wiki/Combined_Statistical_Area}.
Also, the following map from Wikipedia is excellent to
visualize the areas:
\url{http://upload.wikimedia.org/wikipedia/commons/7/7b/Combined_statistical_areas_of_the_United_States_and_Puerto_Rico.gif}
}
\keyword{datasets}

