\docType{data}
\name{corebased_areas}
\alias{corebased_areas}
\title{Core-based Statistical Area (CBSAs)}
\format{A data frame with 917 rows and 4 variables.}
\usage{
data(corebased_areas)
}
\description{
The U.S. Census Bureau groups counties into CBSAs primarily
based on county population. NOTE: Not all counties are
members of a CBSA. For a detailed description, Wikipedia
has an excellent discussion:
\url{http://en.wikipedia.org/wiki/Core_Based_Statistical_Area}.
Also, the following map from Wikipedia is excellent to
visualize the areas:
\url{http://upload.wikimedia.org/wikipedia/commons/7/7b/Combined_statistical_areas_of_the_United_States_and_Puerto_Rico.gif}
}
\keyword{datasets}

