% Generated by roxygen2 (4.0.1): do not edit by hand
\name{explore_counties}
\alias{explore_counties}
\title{Explore U.S. Census Quick Facts}
\usage{
explore_counties()
}
\description{
This function launches a Shiny app for quick visualization of the
\code{quick_facts} dataset. The Shiny, dplyr, leaflet, and RColorBrewer
libraries must be installed.
}
\examples{
explore_counties()
}
\seealso{
\code{\link{quick_facts}} for a description of the dataset
}

