\docType{package}
\name{noncensus}
\alias{noncensus}
\alias{noncensus-package}
\alias{package-noncensus}
\title{U.S. Census Region and Demographic Data}
\description{
The R package \code{noncensus} provides a collection of
various regional information determined by the U.S. Census
Bureau along with demographic data. We have also included
scripts to download, process, and load the original data
from their sources.
}

