% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_choro}
\alias{plot_choro}
\title{Plot a Choropleth}
\usage{
plot_choro(df, fill, map = c("county", "state", "world"), palette = "Blues",
  cuts = NULL, continental = T)
}
\arguments{
\item{df}{The dataframe with column "fips" (the FIPS
code of the counties or states to show), with a column for the data to show}

\item{fill}{The name of the variable to show in the choropleth}

\item{map}{The level at which to draw the map. Options are "county", "state",
"world"}

\item{palette}{An RColorBrewer palette to use. Default is "Blues"}

\item{cuts}{An optional vector specifying where to make the color breaks
Default cuts are the 20th, 40th, 60th, and 80th percentiles}

\item{continental}{A logical whether to show the continental US}
}
\description{
This function takes a dataframe and generates a choropleth plot
}
\examples{
data(population_age, package = "noncensus")
df <- plyr::ddply(population_age, "fips", summarize, population = sum(population))
plot_choro(df, fill = "population", map = "county", palette = "Purples",
continental = TRUE)
}

