% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{population_age}
\alias{population_age}
\title{County Population Data by Age}
\format{A data frame with 56574 rows and 4 variables}
\usage{
data(population_age)
}
\description{
A dataset containing the population totals and percentages for each US county
by age of inhabitant. The variables are as follows:
}
\details{
\itemize{
  \item fips. FIPS code for the county
  \item age_group. Age groups are in 5 year intervals
  \item population. Count of people in that county in that age group
  \item percent. Percent of that county's total population in that age group
}
}
\keyword{datasets}

