% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{state_polygons}
\alias{state_polygons}
\title{Polygons to Describe Each U.S. State's Geographical Shape}
\format{A data frame with 26,819 rows and 6 variables}
\usage{
data(state_polygons)
}
\description{
Data containing the vertices to describe each U.S. state's geographical
shape as a polygon. Vertices are given in terms of latitude and
longitude. The order in which the vertices should be drawn are also given by
\code{order}.
}
\details{
NOTE: A small number of states have multiple groups. Typically, these
counties are separated into multiple polygons by, say, a body of water.
Example: Texas (FIPS == 48).

U.S. states are uniquely identified by a FIPS code.

The \code{state_polygons} data frame consists of the following variables:
\itemize{
  \item order. The order in which the vertices should be drawn.
  \item long. States's longitude
  \item lat. States's latitude
  \item group. State polygon's group number.
  \item state. The states's name.
  \item fips. State FIPS Code
}
}
\keyword{datasets}

