\docType{data}
\name{states}
\alias{states}
\title{Demographic Data and Census Regions of U.S. States}
\format{A data frame with 51 rows and 3 variables}
\usage{
data(states)
}
\description{
A dataset containing demographic information and the census
regions of each U.S. state as defined by the U.S. Census
Bureau. The U.S. is divided into four regions:
}
\details{
\enumerate{ \item Midwest \item Northeast \item South \item
West }

Within each region, states are further partitioned into
divisions. This data frame provides the \code{region} and
\code{division} for each U.S. State.

For more details about census regions, see:
\url{http://en.wikipedia.org/wiki/List_of_regions_of_the_United_States#Census_Bureau-designated_regions_and_divisions}
}
\keyword{datasets}

