% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{world_polygons}
\alias{world_polygons}
\title{Polygons to Describe Each Country's Geographical Shape}
\format{A data frame with 406,940 rows and 11 variables}
\usage{
data(world_polygons)
}
\description{
Data containing the vertices to describe each country's geographical
shape as a polygon. Vertices are given in terms of latitude and
longitude. The order in which the vertices should be drawn are also given by
\code{order}.
}
\details{
NOTE: Many countries have multiple groups. Typically, these
countries are separated into multiple polygons by, say, a body of water.
Example: United States (FIPS == US, UN code == 840).

Countries are uniquely identified in many ways: by a FIPS code, ISO2, IS03,
and by the UN code.

The \code{world_polygons} data frame consists of the following variables:
\itemize{
  \item order. The order in which the vertices should be drawn
  \item long. Country's longitude
  \item lat. Country's latitude
  \item group. Country polygon's group number
  \item name. Unique name to describe country
  \item fips. Country FIPS Code
  \item ISO2. Country ISO2 Code
  \item ISO3. Country ISO3 Code
  \item region. Country region (continent)
  \item subregion. Country subregion
  \item UN_code. Country UN Code
}
}
\keyword{datasets}

