% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_choro.r
\name{shiny_choro}
\alias{shiny_choro}
\title{Start a Shiny Application for Choropleths}
\usage{
shiny_choro(df, fill, categories = NULL, map = c("county", "state",
  "world"), palette = "Blues", background = c("Base", "Greyscale",
  "Physical", "None"), cuts = NULL, dir = NULL)
}
\arguments{
\item{df}{The dataframe with column \code{fips} (the FIPS code of the
counties or states to show), with a column for the data to show, and a column
for the grouping variable (if specifying categories)}

\item{fill}{The name of the variable to show in the choropleth}

\item{categories}{The name of the (optional) grouping variable on which to
divide the data}

\item{map}{The level at which to draw the map. Options are \code{county},
\code{state}, and \code{world}}

\item{palette}{An RColorBrewer palette to use. Default is \code{Blues}}

\item{background}{One of \code{Base}, \code{Greyscale}, \code{Physical}, or
\code{None}, to have as the background tiles for the map}

\item{cuts}{An optional vector specifying where to make the color breaks.
Default cuts are the 20th, 40th, 60th, and 80th percentiles}

\item{dir}{The directory in which to create the Shiny app. Defaults to
\code{\link[base]{tempdir}}}
}
\description{
This function takes a dataframe and generates a local Shiny application for
visualizing the results in a choropleth.
}
\details{
NOTE: The \code{leaflet} R package must be installed.
}
\examples{
data(population_age, package = "noncensus")
shiny_choro(population_age, fill = "population", categories = "age_group",
            map = "county", palette = "Purples", background = "Grey")
}

